requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.0.0
flask[async]>=3.0.0
pydantic>=2.5.2
pytest>=7.4.3
